Fix multithreaded requests to LazyLeastSquares
I expected @lazy to handle this, but evidently it does not:
if multiple threads attempted to fetch frames concurrently,
the body of `def frames` could be executed in each thread.
This led to an error on `del self.op`, because another thread
had already deleted it. But even setting the error aside,
this is just a waste of resources, which the laziness was
intended to prevent in the first place.

Instead of @lazy, explicitly use a lock to ensure mutual exclusion.